Correccion en el tipo de dato

+++ vucem-catalogos-service/src/main/java/com/example/vucem_catalogos_service/model/dto/CatUnidadAdministrativaDTO.java
@@ -21,7 +21,7 @@ public class CatUnidadAdministrativaDTO {
     private String descripcion;
     private String cveEntidad;
     private String nombreEntidad;
-    private Short idDependencia;
+    private Long idDependencia;
     private String nombreDependencia;
     private Boolean blnFronteriza;
     private Instant fecIniVigencia;
