+++ vucem-catalogos-service/src/main/java/com/example/vucem_catalogos_service/persistence/repo/ICatDependenciaRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ICatDependenciaRepository extends JpaRepository<CatDependencia, Short> {
+public interface ICatDependenciaRepository extends JpaRepository<CatDependencia, Long> {
 
 
     @Query("""
